api/requirements.txt
anyio==3.6.2
bcrypt==4.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
ecdsa==0.18.0
fastapi==0.89.1
google-auth==2.16.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
gspread==5.7.2
h11==0.14.0
httptools==0.5.0
idna==3.4
numpy==1.24.2
oauthlib==3.2.2
pandas==1.5.3
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.2
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
